Support for `i18n` warning
The following warning appears in i18n, but it seems to be OK, so the part that was temporarily supported is removed.

```
You are running the esm-bundler build of vue-i18n. It is recommended to configure your bundler to explicitly replace feature flag globals with boolean literals to get proper tree-shaking in the final bundle.
```

https://github.com/intlify/vue-i18n-next/issues/810#issuecomment-979248446
> In dev mode (vite dev), This warning is intended.

refs: ["VUE3-BP-T-5"]